requirements.txt
# Core Libraries
ccxt>=4.0.0          # 거래소 API (Pro 버전 포함)
pandas>=2.0.0         # 데이터 처리
numpy>=1.24.0         # 수치 연산
pydantic>=2.0.0       # 데이터 검증 (Python 3.9+ 호환)
pyupbit>=0.2.33       # 업비트 API (실거래)
aiofiles>=22.0.0      # 비동기 파일 처리 (Python 3.9+ 호환)

# ML/DL Libraries (Phase 2-3)
torch>=2.0.0          # PyTorch (LSTM)
scikit-learn>=1.3.0   # ML utilities
xgboost>=1.7.0        # XGBoost (compatible with Python 3.9)
gymnasium>=0.29.0     # RL environment (OpenAI Gym successor)
stable-baselines3>=2.0.0  # RL (PPO/DQN)
tensorboard>=2.10.0   # Training visualization for RL
optuna>=3.0.0         # 하이퍼파라미터 튜닝
shap>=0.43.0          # Model explainability
ta>=0.11.0            # Technical indicators
matplotlib>=3.8.0     # Visualization
seaborn>=0.13.0       # Statistical plots
plotly>=5.0.0         # Interactive visualizations
streamlit>=1.0.0      # Dashboard framework
jinja2>=3.0.0         # Template engine for reports
scipy>=1.0.0          # Scientific computing
statsmodels>=0.14.0   # Statistical models

# Data & API
python-dotenv>=1.0.0  # 환경변수
cryptography>=41.0.0  # API 키 암호화
requests>=2.31.0      # HTTP 요청
websocket-client>=1.7.0  # WebSocket
aiohttp>=3.9.0        # 비동기 HTTP

# Development Tools
pytest>=7.4.0         # 테스트
pytest-asyncio>=0.21.0  # 비동기 테스트
pytest-cov>=4.1.0     # 커버리지
black>=23.0.0        # 코드 포맷터
isort>=5.12.0         # import 정렬
flake8>=6.0.0         # 린터
mypy>=1.8.0          # 타입 체크
pre-commit>=3.5.0     # Git hooks
bandit>=1.7.0         # 보안 스캔
safety>=2.3.0         # 취약점 체크

# Monitoring
prometheus-client>=0.19.0  # 메트릭
loguru>=0.7.0        # 향상된 로깅

# Database (for Phase 4)
# postgresql==14.0
# sqlalchemy==2.0.25